+++ Backend/CoreService/Core.Application/Features/Trading/Orders/Commands/CreateOrder/CreateOrderHandler.cs
@@ -1,10 +1,8 @@
 ﻿using System.Text.Json;
-using System.Text.Json.Nodes;
 using Core.Application.Configurations;
-using Core.Application.Dtos.Statistics;
 using Core.Application.Dtos.Statistics.Messages;
 using Core.Application.Dtos.Trading;
-using Core.Application.Interfaces;
+using Core.Application.Interfaces.Statistics;
 using Core.Application.Wrappers;
 using Core.Application.Wrappers.Enums;
 using Core.Domain.Entities.Trading;
@@ -22,14 +20,15 @@ public class CreateOrderHandler :
     private readonly IMapper _mapper;
     private readonly IStatisticMessageProducer _messageProducer;
 
-    public CreateOrderHandler(ICartRepository cartRepository, IOrderRepository orderRepository, IMapper mapper, IStatisticMessageProducer messageProducer)
+    public CreateOrderHandler(ICartRepository cartRepository, IOrderRepository orderRepository, IMapper mapper,
+        IStatisticMessageProducer messageProducer)
     {
         _cartRepository = cartRepository;
         _orderRepository = orderRepository;
         _mapper = mapper;
         _messageProducer = messageProducer;
     }
-    
+
 
     public async Task<Result<OrderDto>> Handle(CreateOrderCommand request, CancellationToken cancellationToken)
     {
@@ -41,7 +40,7 @@ public class CreateOrderHandler :
         }
 
         var order = new Order(cart, request.UserId);
-        
+
         order.SetOrderItem(CreateOrderItems(order, cart.CartItems));
         var newOrder = await _orderRepository.AddAsync(order, cancellationToken);
 
@@ -50,24 +49,25 @@ public class CreateOrderHandler :
         {
             return Result<OrderDto>.Failed(ErrorTypeCode.EntityConflict);
         }
-     
-        
+
+
         var data = _mapper.Map<OrderDto>(newOrder);
 
-      
-        var userStatisticMessage = new UserStatisticMessage() {UserId = request.UserId, OrderCreated = 1, TotalRevenue = (int)data.TotalAmount };
+
+        var userStatisticMessage = new UserStatisticMessage()
+            { UserId = request.UserId, OrderCreated = 1, TotalRevenue = (ulong)data.TotalAmount };
         var userJsonMessage = JsonSerializer.Serialize(userStatisticMessage, JsonSerializerOptions.Default);
-        
-        await _messageProducer.SendMessageAsync(StatisticBrokerConfiguration.UserTopic, userJsonMessage, cancellationToken);
-        
+
+        await _messageProducer.SendMessageAsync(StatisticBrokerConfiguration.UserTopic, userJsonMessage,
+            cancellationToken);
+
         return Result<OrderDto>.Successful(data);
     }
 
     private IList<OrderItem> CreateOrderItems(Order order, ICollection<CartItem> cartItems)
     {
-       
         var items = cartItems.Select(x =>
-            new OrderItem(order, x.ProductObject, x.Quantity, x.Quantity * x.ProductObject.Price))
+                new OrderItem(order, x.ProductObject, x.Quantity, x.Quantity * x.ProductObject.Price))
             .ToList();
 
         return items;
